+++ CHANGES.md
@@ -1,5 +1,5 @@
 
-
+- Add `Uuidm.{to,of}_mixed_endian_bytes`.
 - Add `Uuidm.unsafe_of_bytes`
 
 v0.9.6 2016-08-12 Zagreb